backend/requirements.txt
flask
python-dotenv
psycopg2-binary
requests
PyPDF2
flask-cors
